rename ahat stream
